update.sh: Add verbose switch
It is annoying to see those unninteresting lines, where the files have
not changed.  Now there will be a switch for that.

+++ update.sh
@@ -1,33 +1,27 @@
-#!/bin/bash
+#! /bin/bash
 #
-# Copied over from Uli Schlachter (psychon) and adjusted until it suited my needs.
-# Written by Ignas Anikevicius and the code is licensed under GPL v3.
+# This is the file which is used for synching the directories
+#
 
-# For every file and dir in ., this scripts checks if a file with .<name>
-# exists in ~ and copies it into the current directory
+print_help() {
+    cat <<EOF
+This is the help file for the script.
 
-DIR_TO=
-DIR_FROM=~/.
+Usage:
+    -b|--backup
+        Update the repository, or backup.
+    -s|--sync
+        Update the local files using the repository copies.
+    -p|--pretend
+        Do not actually sync, just check if there are any changes.
+    -v|--verbose
+        A lot of output which is not needed always.
+    -h|--help
+        Print this text.
+EOF
+}
 
-if [[ -z $1 ]]; then
-    utype="repo"
-elif [[ "$1" == "repo" || "$1" == "local" ]];then
-    utype="$1"
-fi
-if [[ "$utype" != "repo" ]]; then
-    local DIR_TMP="$DIR_FROM"
-    DIR_FROM="$DIR_TO"
-    DIR_TO="$DIR_TMP"
-fi
-
-high_skip="\033[1;33m"
-high_unknown="\033[1;37m"
-high_update="\033[1;34m"
-high_outdate="\033[1;31m"
-high_end="\033[0;37m"
-
-update ()
-{
+update () {
     file=${DIR_TO}$@
     base_file=${DIR_FROM}$file
 
@@ -35,17 +29,27 @@ update ()
     then
         if diff "$base_file" "$file" > /dev/null; then
             # File did not change
-            #echo -e "${high_skip}Skipping${high_end} $file (no changes)"
+            if ${VERBOSE}; then
+                echo -e "${col_skip}Skipping${col_end} $file"
+            fi
             return
         else
             if [[ `stat -c %Y $base_file` -lt `stat -c %Y $file` ]]; then
-                echo -e "${high_outdate}Skipping${high_end} $file"
-                echo -e "        (local copy is different and older than the repo copy)"
+                echo -e "${col_outdate}Skipping${col_end} $file"
                 return
             fi
             # Normal files are just copied
-            echo -e "${high_update}Updating${high_end} $file" >& 2
-            cp -R "$base_file" "$file"
+            if ${SYNC}; then
+                echo -e "${col_update}Synchronising${col_end} $base_file" >& 2
+                if ! ${PRETEND}; then
+                    cp -vira "$file" "$base_file"
+                fi
+            else
+                echo -e "${col_update}Updating${col_end} $file" >& 2
+                if ! ${PRETEND}; then
+                    cp -R "$base_file" "$file"
+                fi
+            fi
             return
         fi
     fi
@@ -58,26 +62,77 @@ update ()
         return
     fi
 
-    echo -e "${high_unknown}Skipping${high_end} $file (unknown type)"
+    echo -e "${col_unknown}Skipping${col_end} $file"
 }
 
 
 update_dir ()
 {
-    for file in $@*
-    do
+    for file in $@*; do
         case $file in
-            update.sh|README|Rakefile)
+            update.sh|README)
                 continue
                 ;;
+            *)
+                update ${file}
+                ;;
         esac
-
-        update "$file"
     done
 }
 
-update_dir
+main() {
+    export PRETEND=false
+    export VERBOSE=false
+    export SYNC=false
+    export DIR_TO=
+    export DIR_FROM=~/.
 
-echo -e "\033[1mEverything is up to date\033[0m"
+    export col_skip="\033[1;33m"
+    export col_unknown="\033[1;37m"
+    export col_update="\033[1;34m"
+    export col_outdate="\033[1;31m"
+    export col_end="\033[0;37m"
 
-# vim: tw=80
+    echo -e "\
+${col_end}\
+============================================================================
+Colour codings:
+${col_update}Updating: Local ver. is older
+${col_skip}Skipping: Local ver. is the same
+${col_outdate}Skipping: Local ver. is newer
+${col_unknown}Skipping: Unknown file type or there is no local copy pressent
+${col_end}\
+============================================================================\
+"
+
+    for var in $@; do
+        case $var in
+            -p|--pretend)
+                export PRETEND=true; shift
+                echo -e "\
+Running in PRETEND mode (no actual changes will be done)
+============================================================================\
+"
+                ;;
+            -v|--verbose)
+                export VERBOSE=true; shift
+                ;;
+            -b|--backup)
+                export SYNC=false
+                update_dir
+                return 0
+                ;;
+            -s|--sync)
+                export SYNC=true
+                update_dir
+                return 0
+                ;;
+            -h|--help|*)
+                print_help 
+                return 0
+                ;;
+        esac
+    done
+}
+
+main $@
